+++ src/com/businessinsights/view/Utils.java
@@ -1,13 +1,9 @@
-package com.businessinsights.ui;
+package com.businessinsights.view;
 
 import java.awt.Toolkit;
 import java.awt.Component;
 import java.awt.Dimension;
 
-/**
- * Ideally this class should in a package under com.fleetrental.view
- *
- */
 public class Utils {
 
     /**
@@ -27,21 +23,4 @@ public class Utils {
         comp.setLocation( (screenSize.width - frameSize.width) / 2,
                 (screenSize.height - frameSize.height) / 2);
     }
-
-    /**
-     * API to show error message
-     * @param e Exception
-     */
-    public static void showErrorMessage(Exception e) {
-        //exception occurred, so do not process further and return now after showing message.
-        MessageDialog msgDlg = new MessageDialog("Error",
-                e.getMessage() != null ?
-                        e.getMessage() : "Error occurred");
-        Utils.centerWindow(msgDlg);
-        e.printStackTrace();
-        msgDlg.setModal(true);
-        msgDlg.show();
-
-    }
-
 }
